+++ common-ux/components/test.spec.ts
@@ -1,0 +1,3 @@
+/// <reference path='../typings/tsd.d.ts' />
+
+// this is used to pull in typings for tests